pyproject.toml
[project]
name = "pluGET"
version = "1.8.2"
authors = [
    { name = "Neocky" }
]
description = "A powerful package manager that updates plugins and server software for Minecraft servers."
readme = "README.md"
license = "Apache-2.0"
license-files = ["LICENSE"]
requires-python = ">=3.10"
dependencies = [
    "requests>=2.25.1",
    "rich>=9.13.0",
    "ruamel-yaml>=0.17.21"
]

[project.optional-dependencies]
dev = [
    "mypy>=1.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "types-requests>=2.28.0"
]

[project.urls]
Repository = "https://github.com/Neocky/pluGET"
Issues = "https://github.com/Neocky/pluGET/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pluget"]
